requirements.txt
torch
torchvision
gradio
litellm
python-dotenv
pydantic-settings
loguru
numpy
pillow
pandas